Reduced delay for delaying album Fixed an issue where DocumentFileCompat would cause a crash on API 26 Fixed an issue where a folder with no name or protected folders like Downloads folder would cause a crash
